Change method to fetch video file
In the old method, a valid cf token is mandatory during the process in order to get a proper video link. However, the way to get a cf token is flaky.

In the new method, instead of targeting the video link, it targets m3u8 playlist. First, download all segments from m3u8 playlist, and then merge all segments to final video file.