dl: implement parallel operations for parsing m3u8 playlist
Before we used FFmpeg to download the m3u8 playlist although
this worked flawlessly it took a long time for the whole thing
to download as FFmpeg downloads each segment individually uses
FFmepg's crypto module to decode it and finally parse it into
a video buffer and this whole process was done on a single
thread and the presence of multiple segments made the whole
process time-consuming.

To overcome this issue I wrote an implementation which
downloads the m3u8 playlist, extract the URLs for the video
segments and the key file (yes, these segments are encrypted)
and then we download the segments and the key file parallelly,
extract key ID from the key file which we use with OpenSSL to
decrypt the video segment once again parallelly and finally
concatenate the decoded video segments and parse them into
FFmpeg.

Without parallel operations:

 - $ time bash a -a "Kill la kill"  -e 1
   [INFO] Downloading Episode 1...
   real    0m35.923s
   user    0m6.954s
   sys     0m1.662s

With parallel operations:

 - $ time bash a -a "Kill la kill"  -e 1 -t 10
   [INFO] Downloading Episode 1...
   real    0m10.886s
   user    0m6.588s
   sys     0m2.924s

Thanks to Akianonymous for the betterment and implementing
my original script into the program.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>