Fix xargs options conflict warning
From xargs man: The options --replace (-I, -i) and --max-args (-n)
are mutually exclusive. If some of them are specified at the same time,
then xargs will generally use the option specified last on the command
line, i.e., it will reset the value of the offending option (given
before) to its default value. Additionally, xargs will issue a warning
diagnostic on stderr. The exception to this rule is that the special
max-args value 1 ('-n1') is ignored after the --replace option and its
aliases -I and -i, because it would not actually conflict.